Move each code section to a respective module inside lib, keep things organized
