+++ python/FACTOR.py
@@ -1,36 +1,27 @@
-import ctypes
-import urllib.request
-import urllib.error
-import urllib.parse
 import base64
 import json
-import hashlib
 import struct
 import random
-import os
 import sys
 import struct
-#import sympy as sp
 import gmpy2 as gmp
 from math import floor, ceil
 from gmpy2 import mpz,mpq,mpfr,mpc
 from gmpy2 import is_square, isqrt, sqrt, log2, gcd, is_prime, next_prime
-from numpy.ctypeslib import ndpointer
 import subprocess
-import random
 from time import time
 import multiprocessing
-import hashlib
-import base58
-from number_theory import *
+from lib.number_theory import *
+from lib.ctypes import *
+from lib.rpc import *
+from lib.repr import *
+from lib.transaction import *
+from lib.factordb_connector import *
+
 #Attibution note: the Bitcoin RPC components here are taken
 #                 from publicly available sources and are not
 #                 my own original code.
 
-RPC_PORT     = os.environ.get("RPC_PORT", "8332")
-RPC_URL      = os.environ.get("RPC_URL", "http://127.0.0.1:"+ str(RPC_PORT) )
-RPC_USER     = os.environ.get("RPC_USER", "rpcuser")
-RPC_PASS     = os.environ.get("RPC_PASS", "rpcpass") 
 SCRIPTPUBKEY = os.environ.get("SCRIPTPUBKEY") 
 
 SIEVE_MAX_LEVEL = os.environ.get("SIEVE_MAX_LEVEL")
@@ -54,299 +45,6 @@ if ( len(SCRIPTPUBKEY) != 44 ):
     print("Your scriptPubKey ("+str(SCRIPTPUBKEY)+") is not a standard 22 byte scriptPubKey.")
     print("Please check and verify this is correct. Your mining rewards will be lost if you proceed with scriptPubKey address.")
     exit(2)
-
-################################################################################
-# CTypes and utility functions
-################################################################################
-class CParams(ctypes.Structure):
-    _fields_=[("hashRounds",ctypes.c_uint32 ),
-              ("MillerRabinRounds",ctypes.c_uint32 )  
-             ]
-    
-class uint1024(ctypes.Structure):
-    _fields_=[("data", ctypes.c_uint64 * 16 )]
-
-class uint256(ctypes.Structure):
-    _fields_=[("data", ctypes.c_uint64 * 4 )]
-    
-def uint256ToInt( m ):
-    ans = 0    
-    for idx,a in enumerate(m):
-        ans += a << (idx*64)
-    return ans
-
-def uint1024ToInt( m ):
-    ans = 0    
-
-    if hasattr(m, 'data'):
-        for idx in range(16):
-            ans += m.data[idx] << (idx*64)
-    else:
-        for idx,a in enumerate(m):
-            ans += a << (idx*64)
-    
-    return ans
-
-def IntToUint1024( m ):
-    ans = [0]*16
-    n = int(m)
-    MASK = (1<<64)-1
-    
-    for idx in range(16):
-        ans[idx] = (m >> (idx*64)) & MASK
-    
-    return (ctypes.c_uint64 * 16)(*ans)
-    
-    
-def hashToArray( Hash ):
-    if Hash == 0:
-        return [0,0,0,0]
-    
-    number = int(Hash,16)
-    MASK = (1 << 64) - 1
-    arr = [ ( number >> 64*(jj) )&MASK for jj in range(0, 4) ]    
-    
-    return arr
-
-
-################################################################################
-# Bitcoin Daemon JSON-HTTP RPC
-################################################################################
-
-
-def rpc(method, params=None):
-    """ 
-    Make an RPC call to the Bitcoin Daemon JSON-HTTP server.
-
-    Arguments:
-        method (string): RPC method
-        params: RPC arguments
-
-    Returns:
-        object: RPC response result.
-    """
-
-    rpc_id = random.getrandbits(32)
-    data = json.dumps({"id": rpc_id, "method": method, "params": params}).encode()
-    auth = base64.encodebytes((RPC_USER + ":" + RPC_PASS).encode()).decode().strip()
-
-    request = urllib.request.Request(RPC_URL, data, {"Authorization": "Basic {:s}".format(auth)})
-
-    with urllib.request.urlopen(request) as f:
-        response = json.loads(f.read())
-
-    if response['id'] != rpc_id:
-        raise ValueError("Invalid response id: got {}, expected {:u}".format(response['id'], rpc_id))
-    elif response['error'] is not None:
-        raise ValueError("RPC error: {:s}".format(json.dumps(response['error'])))
-
-    return response['result']
-
-################################################################################
-# Bitcoin Daemon RPC Call Wrappers
-################################################################################
-
-
-def rpc_getblocktemplate():
-    try:
-        return rpc("getblocktemplate", [{"rules": ["segwit"]}])
-    except ValueError:
-        return {}
-
-
-def rpc_submitblock(block_submission):
-    return rpc("submitblock", [block_submission])
-
-
-def rpc_getblockcount():
-    return rpc( "getblockcount" )
-
-
-################################################################################
-# Representation Conversion Utility Functions
-################################################################################
-
-
-def int2lehex(value, width):
-    """
-    Convert an unsigned integer to a little endian ASCII hex string.
-    Args:
-        value (int): value
-        width (int): byte width
-    Returns:
-        string: ASCII hex string
-    """
-
-    return value.to_bytes(width, byteorder='little').hex()
-
-
-def int2varinthex(value):
-    """
-    Convert an unsigned integer to little endian varint ASCII hex string.
-    Args:
-        value (int): value
-    Returns:
-        string: ASCII hex string
-    """
-
-    if value < 0xfd:
-        return int2lehex(value, 1)
-    elif value <= 0xffff:
-        return "fd" + int2lehex(value, 2)
-    elif value <= 0xffffffff:
-        return "fe" + int2lehex(value, 4)
-    else:
-        return "ff" + int2lehex(value, 8)
-
-
-def bitcoinaddress2hash160(addr):
-    """
-    Convert a Base58 Bitcoin address to its Hash-160 ASCII hex string.
-    Args:
-        addr (string): Base58 Bitcoin address
-    Returns:
-        string: Hash-160 ASCII hex string
-    """
-
-    table = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz"
-
-    hash160 = 0
-    addr = addr[::-1]
-    for i, c in enumerate(addr):
-        hash160 += (58 ** i) * table.find(c)
-
-    # Convert number to 50-byte ASCII Hex string
-    hash160 = "{:050x}".format(hash160)
-
-    # Discard 1-byte network byte at beginning and 4-byte checksum at the end
-    return hash160[2:50 - 8]
-
-################################################################################
-# Transaction Coinbase and Hashing Functions
-################################################################################
-
-
-def tx_encode_coinbase_height(height):
-    """
-    Encode the coinbase height, as per BIP 34:
-    https://github.com/bitcoin/bips/blob/master/bip-0034.mediawiki
-    Arguments:
-        height (int): height of the mined block
-    Returns:
-        string: encoded height as an ASCII hex string
-    """
-
-    width = (height.bit_length() + 7 )//8 
-
-    return bytes([width]).hex() + int2lehex(height, width)
-
-def make_P2PKH_from_public_key( publicKey = "03564213318d739994e4d9785bf40eac4edbfa21f0546040ce7e6859778dfce5d4" ):
-    from hashlib import sha256 as sha256
-   
-    address   = sha256( bytes.fromhex( publicKey) ).hexdigest()
-    address   = hashlib.new('ripemd160', bytes.fromhex( address ) ).hexdigest()
-    address   = bytes.fromhex("00" + address)
-    addressCS = sha256(                address     ).hexdigest()
-    addressCS = sha256( bytes.fromhex( addressCS ) ).hexdigest()
-    addressCS = addressCS[:8]
-    address   = address.hex() + addressCS
-    address   = base58.b58encode( bytes.fromhex(address))
-    
-    return address
-    
-def tx_make_coinbase(coinbase_script, pubkey_script, value, height, wit_commitment):
-    """
-    Create a coinbase transaction.
-    Arguments:
-        coinbase_script (string): arbitrary script as an ASCII hex string
-        address (string): Base58 Bitcoin address
-        value (int): coinbase value
-        height (int): mined block height
-    Returns:
-        string: coinbase transaction as an ASCII hex string
-    """
-    # See https://en.bitcoin.it/wiki/Transaction
-    coinbase_script = tx_encode_coinbase_height(height) + coinbase_script
-
-    tx = ""
-    # version
-    tx += "02000000"
-    # in-counter
-    tx += "01"
-    # input[0] prev hash
-    tx += "0" * 64
-    # input[0] prev seqnum
-    tx += "ffffffff"
-    # input[0] script len
-    tx += int2varinthex(len(coinbase_script) // 2)
-    # input[0] script
-    tx += coinbase_script
-    # input[0] seqnum
-    tx += "00000000"
-    # out-counter
-    tx += "02"
-    # output[0] value
-    tx += int2lehex(value, 8)
-    # output[0] script len
-    tx += int2varinthex(len(pubkey_script) // 2)
-    # output[0] script
-    tx += pubkey_script
-    # witness commitment value
-    tx += int2lehex(0, 8)
-    # witness commitment script len
-    tx += int2varinthex(len(wit_commitment) // 2)
-    # witness commitment script
-    tx += wit_commitment
-    # lock-time
-    tx += "00000000"
-
-    return tx
-
-
-def tx_compute_hash(tx):
-    """
-    Compute the SHA256 double hash of a transaction.
-    Arguments:
-        tx (string): transaction data as an ASCII hex string
-    Return:
-        string: transaction hash as an ASCII hex string
-    """
-
-    return hashlib.sha256(hashlib.sha256(bytes.fromhex(tx)).digest()).digest()[::-1].hex()
-
-
-def tx_compute_merkle_root(tx_hashes):
-    """
-    Compute the Merkle Root of a list of transaction hashes.
-    Arguments:
-        tx_hashes (list): list of transaction hashes as ASCII hex strings
-    Returns:
-        string: merkle root as a big endian ASCII hex string
-    """
-    
-    # Convert list of ASCII hex transaction hashes into bytes
-    tx_hashes = [bytes.fromhex(tx_hash)[::-1] for tx_hash in tx_hashes]
-
-    # Iteratively compute the merkle root hash
-    while len(tx_hashes) > 1:
-        # Duplicate last hash if the list is odd
-        if len(tx_hashes) % 2 != 0:
-            tx_hashes.append(tx_hashes[-1])
-
-        tx_hashes_new = []
-
-        for i in range(len(tx_hashes) // 2):
-            # Concatenate the next two
-            concat = tx_hashes.pop(0) + tx_hashes.pop(0)
-            # Hash them
-            concat_hash = hashlib.sha256(hashlib.sha256(concat).digest()).digest()
-            # Add them to our working list
-            tx_hashes_new.append(concat_hash)
-
-        tx_hashes = tx_hashes_new
-
-    # Format the root in big endian ascii hex
-    return tx_hashes[0][::-1].hex()
 
 Count = 0
 ################################################################################
@@ -657,8 +355,6 @@ def getParams():
     param.MillerRabinRounds = 50
     return param
 
-gHash = ctypes.CDLL("../libs/gHash.so").gHash
-gHash.restype = uint1024
 
 def mine():
     global SCRIPTPUBKEY
